% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPtoData.R
\name{coeffRV}
\alias{coeffRV}
\title{Escoufier's RV coeff, measure of similarity}
\usage{
coeffRV(X, Y)
}
\arguments{
\item{X}{Numeric Matrix}

\item{Y}{Numeric Matrix}
}
\description{
Escoufier's RV coeff, measure of similarity
}
\details{
This is an implementation of Escoufier's RV coefficient.
The memory needed is of order \eqn{p^2} with \eqn{p} the number of variables.
}

