% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cors}
\alias{Cors}
\title{Observed spearman correlations in X1 dataset}
\format{A positive (semi)definite correlation matrix}
\source{
Korcula and Vis data
}
\usage{
Cors
}
\description{
Observed spearman correlations in X1 dataset
}
\keyword{datasets}
