% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cors2}
\alias{Cors2}
\title{Observed spearman correlations in X2 dataset}
\format{A positive (semi)definite correlation matrix}
\source{
Korcula and Vis data
}
\usage{
Cors2
}
\description{
Observed spearman correlations in X2 dataset
}
\keyword{datasets}
