% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPtoData.R
\name{GenData}
\alias{GenData}
\title{Generate simulated omics data}
\usage{
GenData(nr_X2vars = 100, similarity_data = 1, within_corr = 0.1,
  weight_shared_SNPs = 0, meas_add_X1 = c(0.25, 0.5),
  meas_mult_X1 = meas_add_X1/5, meas_add_X2 = meas_add_X1/2,
  meas_mult_X2 = meas_mult_X1/5)
}
\arguments{
\item{nr_X2vars}{Number of X2 variables}

\item{similarity_data}{Between 0 and 1. How similar are X1 and X2?}

\item{within_corr}{Between 0 and 1. Within correlation of X2.}

\item{weight_shared_SNPs}{Larger or equal to zero. Obsolete!}

\item{meas_add_X1}{A vector of two numerics.}

\item{meas_mult_X1}{A vector of two numerics.}

\item{meas_add_X2}{A vector of two numerics.}

\item{meas_mult_X2}{A vector of two numerics.}
}
\value{
A list with the data:
\item{X1}{The first dataset without measurement error.}
\item{X1m}{The first dataset with measurement error.}
\item{X2}{The second dataset without measurement error.}
\item{X2m}{The first dataset with measurement error.}
\item{y}{Continuous outcome}
\item{d1}{Binary outcome ...}
\item{d}{Binary outcome ...}
}
\description{
Generate simulated omics data using genetics from the MIMOmics consortium.
}
\details{
Number of X2 variables
p_X2 = 100
Set a proportion between 0 and 1 (corresponding to RV coeff between 0 and 0.98, depends on the weight_shared_SNPs)
This is most influential on the cor between X1 and X2, determines to which extent variables in X1 and X2 agree
similarity_data = 1
# Set within correlation in X2 by changing proportion of SNPs used in each X2 variable.
within_corr = 0.1
# Set weight of the shared SNPs for X2. More weight means shared SNPs are more important for X2
#       However as the linear combinations in X1 are not the same as the linear combinations in X2,
#       increasing the weights does not necessarily increase the similarity
#       Set between 0 and Inf, 0 gives approx independent X1 and X2 if similarity_data = 0
weight_shared_SNPs = 0
# Set Measurement error correlations and variances for X1 (first element cor, then var)
meas_add_X1 = c(0.25, 0.5)
meas_mult_X1 = c(0.05, 0.1)
meas_add_X2 = c(0.5, 1) / 4
meas_mult_X2 = c(0.01, 0.02)
}
