% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPtoData.R
\docType{package}
\name{MIMOmicsData}
\alias{MIMOmicsData}
\alias{MIMOmicsData-package}
\title{Generate Omics simulated data}
\description{
Generate Omics simulated data based on Glycan genetics and Measurement error.
}
\section{To Do}{

Age sex, but which scale and for which glycans in what order?
Simulate genotypes, but how about correlation structure?
}

\section{Background}{

One.
}

\section{Overview}{

Two.
}

\section{Details}{

Three.
}

\author{
MIMOmics consortium

Maintainer
MIMOmics (\email{mimomics@lumc.nl})
}
\keyword{DataPackage}
\keyword{MIMOmics}
\keyword{Omics}
