% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.MIMOmicsData}
\alias{plot.MIMOmicsData}
\title{Plot heatmaps of simulated data}
\usage{
\method{plot}{MIMOmicsData}(x, ...)
}
\arguments{
\item{x}{A MIMOmicsData object as generated by \code{\link{GenData}}.}
}
\description{
Plot heatmaps of simulated data
}
\keyword{internal}
