% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.SimuData}
\alias{plot.SimuData}
\title{Plot difference of Simulated and Observed correlations}
\usage{
\method{plot}{SimuData}(x, Corr = Cors, ...)
}
\arguments{
\item{x}{SimuData dataset}
}
\description{
Plot difference of Simulated and Observed correlations
}
