% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.MIMOmicsData}
\alias{print.MIMOmicsData}
\title{Print RV coefs of simulated data}
\usage{
\method{print}{MIMOmicsData}(x, digits = 3, ...)
}
\arguments{
\item{x}{A MIMOmicsData object as generated by \code{\link{GenData}}.}
}
\description{
Print RV coefs of simulated data
}
\keyword{internal}
