% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Probs2}
\alias{Probs2}
\title{Observed allele proportions in X2 dataset}
\format{A matrix with SNPs as columns and genotype as rows}
\source{
Korcula and Vis data
}
\usage{
Probs2
}
\description{
Observed allele proportions in X2 dataset coded as 0, 1 and 2
}
\keyword{datasets}

