% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPtoData.R
\name{SimuGeno}
\alias{SimuGeno}
\title{Simulate genetic data from observed frequencies and correlations}
\usage{
SimuGeno(nr_samples = 1742, Pr = Probs, Corr = Cors)
}
\arguments{
\item{nr_samples}{Integer, sample size}

\item{Pr}{Numeric matrix with frequencies, one of datasets in this package}

\item{Corr}{Numeric matrix with correlations, one of datasets in this package}
}
\description{
Simulate genetic data from observed frequencies and correlations
}

