% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPtoData.R
\docType{package}
\name{MIMOmicsData}
\alias{MIMOmicsData}
\alias{MIMOmicsData-package}
\title{Generate Omics simulated data}
\description{
Generate Omics simulated data based on Glycan genetics and Measurement error.
}
\section{To Do}{

\itemize{
\item{}{Age sex, but which scale and for which glycans in what order?}
\item{}{Allow for nr X2 variables smaller than 73.}
\item{}{Allow for arbitraty nr of X1 variables.}
\item{}{Nr of samples.}
}
}

\section{Background}{

One.
}

\section{Overview}{

Two.
}

\section{Details}{

Three.
}

\author{
MIMOmics consortium

Maintainer
MIMOmics (\email{mimomics@lumc.nl})
}
\keyword{DataPackage}
\keyword{MIMOmics}
\keyword{Omics}
